refactor(schema): move schema module to the core module
