+++ packages/devqubit-engine/src/devqubit_engine/core/schema/devqubit.envelope.1.0.schema.json
